backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv==1.2.2
psycopg2-binary==2.9.11
groq
requests==2.33.0
google-genai
yt-dlp
